Increase time to wait before taking screenshot (#314)
* Increase time to wait before taking screenshot

Hopefully this should fix these screenshots coming out blank every once
in a while like caused by the (async) scripts setting everything up not
having reached the point where the canvases are drawn to when the
screenshots are taken.

* CI: Verify screenshots to be larger than 1k bytes

The screenshots sometime come out blank, in which case the generated PNG
is 800-something bytes. Fail the build in that case to prevent
deploying.